in the middle of a rework
finally have a way to talk about the position of celestial objects in an absolute reference frame, rather than only in relation to each other (UniversePosition).
i am also trying to rework how the sky is rendered, and how the system map works so that you can simply move nearby a star system to see its contents.
this is a part of a larger rework to better support space stations, and the ability to fly them around the universe.